+++ metadata.php
@@ -52,7 +52,7 @@ $aModule = [
             </ul>',
     ],
     'thumbnail' => 'logo.svg',
-    'version' => '1.2.0-rc.6',
+    'version' => '1.2.0-rc.7',
     'author' => 'OXID eSales AG',
     'url' => 'https://www.oxid-esales.com',
     'email' => 'info@oxid-esales.com',
@@ -272,152 +272,164 @@ $aModule = [
             'type' => 'str',
             'value' => ''
         ],
-        // unzer invoice keypairs
+
+        // live Paylater Invoice B2C EUR
         [
             'group' => 'unzerinvoice',
-            'name' => 'sandbox-UnzerPublicKeyB2CEUR',
+            'name' => 'production-UnzerPayLaterInvoiceB2CEURPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
             'group' => 'unzerinvoice',
-            'name' => 'sandbox-UnzerPrivateKeyB2CEUR',
+            'name' => 'production-UnzerPayLaterInvoiceB2CEURPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // live Paylater Invoice B2B EUR
+        [
+            'group' => 'unzerinvoice',
+            'name' => 'production-UnzerPayLaterInvoiceB2BEURPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
             'group' => 'unzerinvoice',
-            'name' => 'sandbox-UnzerPublicKeyB2BEUR',
+            'name' => 'production-UnzerPayLaterInvoiceB2BEURPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // live Paylater Invoice B2C CHF
+        [
+            'group' => 'unzerinvoice',
+            'name' => 'production-UnzerPayLaterInvoiceB2CCHFPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
             'group' => 'unzerinvoice',
-            'name' => 'sandbox-UnzerPrivateKeyB2BEUR',
+            'name' => 'production-UnzerPayLaterInvoiceB2CCHFPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // live Paylater Invoice B2B CHF
+        [
+            'group' => 'unzerinvoice',
+            'name' => 'production-UnzerPayLaterInvoiceB2BCHFPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
             'group' => 'unzerinvoice',
-            'name' => 'sandbox-UnzerPublicKeyB2CCHF',
+            'name' => 'production-UnzerPayLaterInvoiceB2BCHFPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // live Paylater Installment B2C EUR
+        [
+            'group' => 'unzerpaylater',
+            'name' => 'production-UnzerPayLaterInstallmentB2CEURPrivateKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        [
+            'group' => 'unzerpaylater',
+            'name' => 'production-UnzerPayLaterInstallmentB2CEURPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // live Paylater Installment B2C CHF
+        [
+            'group' => 'unzerpaylater',
+            'name' => 'production-UnzerPayLaterInstallmentB2CCHFPrivateKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        [
+            'group' => 'unzerpaylater',
+            'name' => 'production-UnzerPayLaterInstallmentB2CCHFPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // sandbox Paylater Invoice B2C EUR
+        [
+            'group' => 'unzerinvoice',
+            'name' => 'sandbox-UnzerPayLaterInvoiceB2CEURPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
             'group' => 'unzerinvoice',
-            'name' => 'sandbox-UnzerPrivateKeyB2CCHF',
+            'name' => 'sandbox-UnzerPayLaterInvoiceB2CEURPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // sandbox Paylater Invoice B2B EUR
+        [
+            'group' => 'unzerinvoice',
+            'name' => 'sandbox-UnzerPayLaterInvoiceB2BEURPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
             'group' => 'unzerinvoice',
-            'name' => 'sandbox-UnzerPublicKeyB2BCHF',
+            'name' => 'sandbox-UnzerPayLaterInvoiceB2BEURPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // sandbox Paylater Invoice B2C CHF
+        [
+            'group' => 'unzerinvoice',
+            'name' => 'sandbox-UnzerPayLaterInvoiceB2CCHFPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
             'group' => 'unzerinvoice',
-            'name' => 'sandbox-UnzerPrivateKeyB2BCHF',
+            'name' => 'sandbox-UnzerPayLaterInvoiceB2CCHFPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // sandbox Paylater Invoice B2B CHF
+        [
+            'group' => 'unzerinvoice',
+            'name' => 'sandbox-UnzerPayLaterInvoiceB2BCHFPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
             'group' => 'unzerinvoice',
-            'name' => 'production-UnzerPublicKeyB2CEUR',
+            'name' => 'sandbox-UnzerPayLaterInvoiceB2BCHFPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // sandbox Paylater Installment B2C EUR
+        [
+            'group' => 'unzerpaylater',
+            'name' => 'sandbox-UnzerPayLaterInstallmentB2CEURPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
-            'group' => 'unzerinvoice',
-            'name' => 'production-UnzerPrivateKeyB2CEUR',
+            'group' => 'unzerpaylater',
+            'name' => 'sandbox-UnzerPayLaterInstallmentB2CEURPublicKey',
+            'type' => 'str',
+            'value' => ''
+        ],
+        // sandbox Paylater Installment B2C CHF
+        [
+            'group' => 'unzerpaylater',
+            'name' => 'sandbox-UnzerPayLaterInstallmentB2CCHFPrivateKey',
             'type' => 'str',
             'value' => ''
         ],
         [
-            'group' => 'unzerinvoice',
-            'name' => 'production-UnzerPublicKeyB2BEUR',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerinvoice',
-            'name' => 'production-UnzerPrivateKeyB2BEUR',
+            'group' => 'unzerpaylater',
+            'name' => 'sandbox-UnzerPayLaterInstallmentB2CCHFPublicKey',
             'type' => 'str',
             'value' => ''
         ],
 
-        [
-            'group' => 'unzerinvoice',
-            'name' => 'production-UnzerPublicKeyB2CCHF',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerinvoice',
-            'name' => 'production-UnzerPrivateKeyB2CCHF',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerinvoice',
-            'name' => 'production-UnzerPublicKeyB2BCHF',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerinvoice',
-            'name' => 'production-UnzerPrivateKeyB2BCHF',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerpaylater',
-            'name' => 'sandbox-UnzerPaylaterPublicKeyB2CEUR',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerpaylater',
-            'name' => 'sandbox-UnzerPaylaterPrivateKeyB2CEUR',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerpaylater',
-            'name' => 'sandbox-UnzerPaylaterPublicKeyB2CCHF',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerpaylater',
-            'name' => 'sandbox-UnzerPaylaterPrivateKeyB2CCHF',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerpaylater',
-            'name' => 'production-UnzerPaylaterPublicKeyB2CEUR',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerpaylater',
-            'name' => 'production-UnzerPaylaterPrivateKeyB2CEUR',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerpaylater',
-            'name' => 'production-UnzerPaylaterPublicKeyB2CCHF',
-            'type' => 'str',
-            'value' => ''
-        ],
-        [
-            'group' => 'unzerpaylater',
-            'name' => 'production-UnzerPaylaterPrivateKeyB2CCHF',
-            'type' => 'str',
-            'value' => ''
-        ],
         [
             'group' => 'unzerother',
             'name' => 'UnzerjQuery',