+++ src/Model/Order.php
@@ -115,7 +115,7 @@ class Order extends Order_parent
         }
 
         $this->initWriteTransactionToDB(
-            $paymentService->getSessionUnzerPayment()
+            $paymentService->getSessionUnzerPayment(true)
         );
 
         // cleanUp Tmp Order
@@ -193,7 +193,7 @@ class Order extends Order_parent
         $tmpOrder = oxNew(TmpOrder::class);
         $tmpOrder->saveTmpOrder($this);
 
-        Registry::getSession()->setVariable('oxOrderIdOfTmpOrder', $this->getId());
+        Registry::getSession()->setVariable('oxOrderIdOfTmpOrder', $orderId);
 
         return $iRet;
     }